This refactor improves code organization and prepares the codebase for future enhancements related to SIP call handling.
