+++ webapp/src/reducers.ts
@@ -14,7 +14,7 @@ import {
 import {reducer as screenSharingIDs} from 'src/state/screen_sharing_ids/reducer';
 import {USER_JOINED, USER_LEFT, USER_REACTED, USER_REACTED_TIMEOUT} from 'src/state/sessions/action_types';
 import {reducer as sessions} from 'src/state/sessions/reducer';
-import {reducer as sipCallDetails} from 'src/state/sip_call_details/reducer';
+import {reducer as sipDetails} from 'src/state/sip_details/reducer';
 import {
     CallsConfigDefault,
     CallsUserPreferences,
@@ -731,7 +731,7 @@ const rootReducer = combineReducers({
     activeCalls,
     hosts,
     screenSharingIDs,
-    sipCallDetails,
+    sipDetails,
     expandedView,
     switchCallModal,
     screenSourceModal,
