+++ webapp/src/state/active_calls/reducer.ts
@@ -1,15 +1,16 @@
 // Copyright (c) 2020-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import {Channel} from '@mattermost/types/channels';
-import {UserThread} from '@mattermost/types/threads';
-import {UserProfile} from '@mattermost/types/users';
-import {Reducer} from 'redux';
+import {type Channel} from '@mattermost/types/channels';
+import {type UserThread} from '@mattermost/types/threads';
+import {type UserProfile} from '@mattermost/types/users';
+import {type Reducer} from 'redux';
 import {CALL_ENDED, UN_INITIALIZED} from 'src/state/common_action_types';
 
 import {ACTIVE_CALL_REGISTERED} from './action_types';
+import {type Actions} from './actions';
 
-type State = {
+export type State = {
     [channelID: string]: {
         callID: string;
         startAt: number;
@@ -21,7 +22,7 @@ type State = {
 
 const emptyState: State = {};
 
-export const reducer: Reducer<State> = (initialState = emptyState, action) => {
+export const reducer: Reducer<State, Actions> = (initialState = emptyState, action) : State => {
     switch (action.type) {
     case UN_INITIALIZED:{
         return emptyState;
