+++ webapp/src/state/sessions/actions.ts
@@ -1,10 +1,10 @@
 // Copyright (c) 2020-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import {Reaction, UserSessionState} from '@mattermost/calls-common/lib/types';
-import {Channel} from '@mattermost/types/channels';
-import {UserProfile} from '@mattermost/types/users';
-import {ActionCallEnded, ActionUnInitialized} from 'src/state/common_actions';
+import {type Reaction, type UserSessionState} from '@mattermost/calls-common/lib/types';
+import {type Channel} from '@mattermost/types/channels';
+import {type UserProfile} from '@mattermost/types/users';
+import {type ActionCallEnded, type ActionUnInitialized} from 'src/state/common_actions';
 
 import {
     SESSIONS_RECEIVED,
