+++ webapp/src/state/sip_details/reducer.ts
